Move MetadataUpdate to top level.
